Make sugar methods instance methods of `browser`
As we are adding more sugar methods it feels inconsistent to have some of them
available as magic globals and some of them attached to `browser`. This commit
moves all of them under `browser`, and renames `waitFor` to `wait` to avoid
conflict with wd's built-in `waitFor` method (which waits on a function
asserter to become true).